♻ Refactor `$BrainFolder` in theBrain scripts
This commit enables to get the value of the `BrainFolder` variable using environmental variables. If the environmental variable is not found, the program sets `BrainFolder` to the default location.

Additionally, an if statement has been added to check if the path exists, and an error message is displayed if the specified path does not exist.