✨ Add `Get-TheBrainDataDirectory.ps1` script
This commit added a new PowerShell script called
`Get-TheBrainDataDirectory.ps1`. The script retrieves the user's
Brain data directory from the `TheBrainMeta.db` database. It imports
the `PSSQLite` module and uses it to query the `MetaSettings` table in
the `TheBrainMeta.db` database. The query retrieves the value of the
`preferences.userbraindatadirectory` setting, which specifies the
location of the user's Brain data directory.

The script also includes error handling to check if the folder exists
and throws an error if it doesn't. If the
`preferences.userbraindatadirectory` setting is not found in the
database, the script defaults to the user's Documents folder.