📄 Add GNU GPLv3 LICENSE file to project root
This commit adds a GNU General Public License version 3 (GPLv3) LICENSE
file to the root directory of the project. This is an important step in
clearly defining the terms under which the software can be used or
modified.

This commit also introduces a GPLv3 license notice in the header
comments of all PowerShell scripts in the project. The addition of the
license notice is in line with our commitment to open source software
and ensures that all users and contributors are aware of the licensing
terms.

The changes include:
- A LICENSE file containing the full text of the GPLv3 has been added
  to the project root.
- A standard GPLv3 license notice has been added to the top of each
  PowerShell (.ps1) file.

By including a GPLv3 LICENSE file, we ensure that the licensing terms
are readily available and clear to all users and contributors. This
promotes transparency and fosters a healthy open-source community.