🐛 fix(theBrain): Exclude backup folder from file search
In this commit, we address the issue where our PowerShell scripts were
incorrectly searching for both original and backup files. The previous
implementation inadvertently included the backup folder in the search
results.

To resolve this, we now use the following way to limit the path(s) to
search:

```powershell
$PathToSearch = Get-ChildItem -Directory -Path $BrainFolder -Exclude ‘Backup’
```

This ensures that only original files are considered, excluding any
files within the 'Backup' folder.

Fixes #16