✨ feat: add parameter to set image width percentage in Resize-TheBrainNotesYouTubeThumbnail.ps1
This commit introduces a new parameter to the `Resize-TheBrainNotesYouTubeThumbnail.ps1` script that allows users to set the width of the image as a percentage. This enhancement provides more flexibility to users by allowing them to specify the desired width of the image.

The changes include:
- Added a new parameter `$NewWidth` to set the width of the image as a percentage.
- Adjusted the string `$width={$CurrentWidth}p` to reflect the new width when `$ImageType` is resized.
- Set 'default' as the default value of the `$ImageType` parameter to simplify the usage of the script.

This change enhances the functionality of the script and makes it more user-friendly by providing more customization options.

Fixes #17